requirements.txt
langchain==0.3.12
langchain-community==0.3.12
langchain-openai==0.2.11
langchain-core==0.3.27
langchain-aws==0.2.1
langchain-google-vertexai==2.0.0
langchain_text_splitters==0.3.3
langchain-weaviate>=0.0.5
boto3==1.34.144
python-dotenv==1.0.1
fastapi==0.115.12
uvicorn==0.28.0
pypdf==4.1.0
unstructured==0.16.11
markdown==3.6
networkx==3.2.1
pandas==2.2.1
openpyxl==3.1.2
docx2txt==0.8
pypandoc==1.13
PyJWT==2.8.0
python-multipart==0.0.19
sentence_transformers==3.1.1
aiofiles==23.2.1
rapidocr-onnxruntime==1.3.24
opencv-python-headless==4.9.0.80
langchain-ollama==0.2.0
langchain-huggingface==0.1.0
cryptography==44.0.1
python-magic==0.4.27
python-pptx==0.6.23
xlrd==2.0.1
pydantic==2.9.2
elysia-ai
weaviate-client>=4.0.0
